requirements.txt
alembic==1.13.1
llama-index==0.10.17
llama-index-readers-file==0.1.8
llama-index-embeddings-openai==0.1.6
llama-index-vector-stores-postgres==0.1.2
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
